Moved Level, fixed some visibility
